Make TimelineView m_stylusPosition private
m_stylusPosition was previously a public variable. There are synchronisation
issues between the timelineview cursor and the ruler diamond. This patch
doesn't fix the syncing issues, but it does create a more controlled
environment so that we can implement the fix.

darcs-hash:20080119112327-80cf8-72c004175aff28961ce99514e91b20b8393efab2